fixing src attribute for logo image
Like you identified, "images/chefu.png" looks for chefu.png in an images folder on the same sublevel as the route we're in. So when we're top level (index, ingredients, users, etc.) it finds it no problem, when we go to a sublevel (ingredients/add, for example), it looks for a file called "add/images/chefu.png". Changing that src attribute to be "/images/chefu.png". The extra slash in front says "go to the top level and find the images folder". You'll make similar mistakes a million times -- I still do this kind of thing all the time!